requirements.txt
functions-framework==3.0.0
google-cloud-pubsub>=2.13.11
firebase-admin>=6.0.1